final upload
github wouldn't allow me to upload the whole thing without deleting a couple of scripts, so I uploaded it in a zip. I re-downloaded it and it has all the files saved.

I don't know the issue, think it has to do with some files being called "main.py" or "main.html", but I'm too lazy to figure out how to fix this and I sure as hell am not going to manually enter in all the files.